Introduce WebSocket Server Handshake handler
