+++ http/org.wso2.carbon.transport.http.netty/src/main/java/org/wso2/carbon/transport/http/netty/listener/HTTPServerChannelInitializer.java
@@ -101,6 +101,9 @@ public class HTTPServerChannelInitializer extends ChannelInitializer<SocketChann
                       new HTTPTraceLoggingHandler("tracelog.http.downstream", LogLevel.DEBUG));
         }
 
+        pipeline.addLast(Constants.WEBSOCKET_SERVER_HANDSHAKE_HANDLER,
+                         new WebSocketServerHandshakeHandler(this.serverConnectorFuture, this.interfaceId));
+
         try {
             pipeline.addLast(Constants.HTTP_SOURCE_HANDLER,
                              new SourceHandler(this.serverConnectorFuture, this.interfaceId));
